[ADD] MySQL 적용
