+++ back/src/main/java/donot/gas/back/BackApplication.java
@@ -2,8 +2,9 @@ package donot.gas.back;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
 
-@SpringBootApplication
+@SpringBootApplication()
 public class BackApplication {
 
 	public static void main(String[] args) {
