+++ back/src/main/java/donot/gas/back/repository/history/HistoryJpaRepository.java
@@ -1,6 +1,8 @@
+/*
 package donot.gas.back.repository.history;
 
 import donot.gas.back.dto.CountDto;
+import donot.gas.back.entity.Order;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
@@ -12,23 +14,6 @@ import java.util.ArrayList;
 public class HistoryJpaRepository {
     private final EntityManager em;
 
-    public CountDto getCountByGrade(Long userId) {
-        ArrayList<Long> cntList = new ArrayList<>();
-        for(int i=1; i<=5; i++) {
-            Long cnt = em.createQuery(
-                    "select COUNT(o.id) from Order o" +
-                            " where o.user.id=:userId and o.grade=:grade", Long.class)
-                    .setParameter("userId", userId)
-                    .setParameter("grade", i)
-                    .getSingleResult();
-            System.out.println("THIS: " + cnt);
-            cntList.add(cnt);
-        }
-        return new CountDto(
-                cntList.get(0),
-                cntList.get(1),
-                cntList.get(2),
-                cntList.get(3),
-                cntList.get(4));
-    }
+
 }
+*/
