+++ back/src/main/resources/application.properties
@@ -1,5 +1,6 @@
-spring.datasource.url = jdbc:h2:tcp://localhost/~/savePolarBear
-spring.datasource.driver-class-name = org.h2.Driver
-spring.datasource.username = sa
-spring.jpa.show-sql = true 
+spring.datasource.driver-class-name = com.mysql.cj.jdbc.Driver
+spring.datasource.url = jdbc:mysql://localhost:3306/dontgas?serverTimezone=UTC&characterEncoding=UTF-8
+spring.datasource.username = dont_gas_admin
+spring.datasource.password = admin1234!
+spring.jpa.show-sql = true
 spring.jpa.hibernate.ddl-auto = none
